+++ deep8/src/math/ReduceSum.cpp
@@ -101,7 +101,7 @@ void ReduceSumCPU(const Tensor &x, Tensor &y, int axis) {
         ReduceSumCPUImpl<float>(device, x.data<float>(), x.shape, y.data<float>, y.shape, axis);
         break;
     case DType::Float64:
-        ReduceSumCPUImpl<float>(device, x.data<float>(), x.shape, y.data<float>, y.shape, axis);
+        ReduceSumCPUImpl<double>(device, x.data<double>(), x.shape, y.data<double>, y.shape, axis);
         break;
     default:
         DEEP8_RUNTIME_ERROR("type " << x.type.name << " is not support");
